+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Nothing yet
 
+## 6.1.4 (2025-03-10)
+
+- Change unavailable resource style [#670](https://github.com/datagouv/udata-front/pull/670)
+
 ## 6.1.3 (2025-03-04)
 
 - Support null end date for temporal coverage [#411](https://github.com/datagouv/udata-front/pull/411)